refactor: altera nomes e pasta onde ficarão os arquivos de usuario para melhor organização
